[Test] 데미지 시스템, 악기변경 작동 테스트

+++ Assets/Student/JWJ/Scripts/OldScripts/Bullet.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class Bullet : MonoBehaviour
 {
 
-    private float damage;
-    public void setDamage(float damageAmount)
+    private int damage;
+    public void setDamage(int damageAmount)
     {
         damage = damageAmount;
     }
